+++ src/Components/Attendees.js
@@ -1,3 +1,61 @@
-export default function Attendees() {
-  return;
+import React, {useState} from "react";
+
+export default function Attendees({attendees, event, updateEventAttendance}) {
+  
+  const [showAttendees, setShowAttendees] = useState(false);
+
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  } 
+  return (
+    <>
+  
+    <button onClick={toggleEventAttendees}>
+      {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+    </button>
+
+    {showAttendees ? (
+      <div className="attendees">
+        {attendees.map((attendee, index) => (
+          <React.Fragment key={index}>
+            <div key={attendee.id} className="attendee">
+              <p>
+                <img
+                  src={attendee.avatar}
+                  alt={attendee.firstName}
+                />
+                {"   "}
+                <span>
+                  {" "}
+                  {attendee.firstName} {attendee.lastName}{" "}
+                </span>
+              </p>
+              <p>
+                <button
+                  className="clickable"
+                  onClick={() =>
+                    updateEventAttendance(
+                      event.id,
+                      attendee.id
+                    )
+                  }
+                >
+                  Attending:
+                </button>
+                <span>
+                  {attendee.attendance ? "✅" : "❌"}
+                </span>
+              </p>
+
+              <p>
+                <span>Note:</span> {attendee.note}
+              </p>
+            </div>
+          </React.Fragment>
+        ))}
+      </div>
+    ) : null}
+  </>
+  )
+
 }
